Renaming health bar constants
