+++ src/js/constants/HealthBarStates.js
